Fixed more compilation warnings.
git-svn-id: http://winbolo.googlecode.com/svn/trunk@67 f5f8f8c0-f940-0410-bb6b-693536db5c03

+++ winbolo/src/gui/win32/dialogSkins.c
@@ -67,8 +67,8 @@ BOOL CALLBACK dialogSkinsCallback( HWND hWnd, unsigned uMsg, WPARAM wParam, LPAR
         char skinName[FILENAME_MAX];
         int nItem, itemNum;
         hList = GetDlgItem(hWnd, IDC_SKINLIST); 
-        nItem = SendMessage(hList, LB_GETCURSEL, 0, 0); 
-        itemNum = SendMessage(hList, LB_GETITEMDATA, nItem, 0); 
+        nItem = (int) SendMessage(hList, LB_GETCURSEL, 0, 0); 
+        itemNum = (int) SendMessage(hList, LB_GETITEMDATA, nItem, 0); 
         SendMessage(hList, LB_GETTEXT, nItem, (LPARAM) (LPCTSTR) skinName); 
         if (strcmp(skinName, DEFAULT_NO_SKIN) == 0) {
           skinsLoadSkin("");
